Move Matcher up a level
